+++ packages/pres/index.js
@@ -1,8 +1,14 @@
-export { Pres, Pres as Preset }                           from './src/Pres.js'
-export { Presm }                                          from './src/Presm.js'
-export { demo, randPres, randPres as randPreset }         from './src/preset-utils.js'
-export { scale, limFF }                                   from './src/color-utils.js'
-export { render }                                         from './src/render.js'
-export { rgaToHsi, hexToHsi, hsiToInt, hsiToHsl, modHsi } from './src/color-bitwise.js'
+export { Pres, Pres as Preset } from './src/Pres.js'
+export { Presm }                from './src/Presm.js'
+export {
+  demo, randPres, randPres as randPreset
+}                               from './src/preset-utils.js'
+export {
+  scale, limFF
+}                               from './src/color-utils.js'
+export { render }               from './src/render.js'
+export {
+  hexToHsi, rgaToHsi, hsaToRgi, hsaToRgi as hsaToInt, hsiToHsl, hsiToRgi, hsiToRgi as hsiToInt, modHsi, modRgi
+}                               from './src/color-bitwise.js'