+++ packages/pres/src/preset-utils.js
@@ -1,26 +1,36 @@
-import { randBetw }         from '@aryth/rand'
-import { intToRgb }         from '@palett/convert'
-import { rgbToStr }         from '@palett/stringify'
-import { hexToHsi, modHsi } from './color-bitwise.js'
-import { Pres }             from './Pres.js'
-import { sequence }         from './sequence.js'
+import { randBetw }                             from '@aryth/rand'
+import { intToRgb }                             from '@palett/convert'
+import { rgbToStr }                             from '@palett/stringify'
+import { hexToHsi, hsiToRgi, modHsi, rgaToHsi } from './color-bitwise.js'
+import { Pres }                                 from './Pres.js'
+import { sequence }                             from './sequence.js'
 
 export function demo(pres, count) {
   return `${sequence.call(pres, count).map(int => rgbToStr(intToRgb(int))).join(' ')} | ${(rgbToStr(intToRgb(pres.nan)))}`
 }
 
-const randUpward = (hsi) => {
+const hsiUpward = (hsi) => {
   const dh = randBetw(-16, 16)
   const ds = randBetw(-8, 8)
   const dl = randBetw(-8, 24)
   return modHsi(hsi, dh, ds, dl)
 }
 
-const randCounter = (hsi) => {
+const hsiReverse = (hsi) => {
   const dh = randBetw(150, 210)
   const ds = randBetw(-32, -24)
   const dl = randBetw(+16, +32)
   return modHsi(hsi, dh, ds, dl)
+}
+
+const rgiReverse = (rgi) => {
+  const r = rgi >> 16 & 0xFF, g = rgi >> 8 & 0xFF, b = rgi >> 0 & 0xFF
+  const hsi0 = rgaToHsi(r, g, b)
+  const dh = randBetw(150, 210)
+  const ds = randBetw(-32, -24)
+  const dl = randBetw(+16, +32)
+  const hsi1 = modHsi(hsi0, dh, ds, dl)
+  return hsiToRgi(hsi1)
 }
 /**
  *
@@ -30,8 +40,8 @@ const randCounter = (hsi) => {
  */
 export const randPres = (hex, name) => {
   const min = hexToHsi(hex)
-  const max = randUpward(min)
-  const nan = randCounter(min)
+  const max = hsiUpward(min)
+  const nan = hsiToRgi(hsiReverse(min))
   const pres = new Pres(min, max, nan)
   if (name?.length) pres.name = name
   return pres