+++ packages/pres/src/sequence.js
@@ -1,5 +1,5 @@
 import { init }         from '@vect/vector-init'
-import { hsaToInt }     from './color-bitwise.js'
+import { hsaToRgi }     from './color-bitwise.js'
 import { limFF, scale } from './color-utils.js'
 
 export function sequence(count = 2) {
@@ -14,5 +14,5 @@ export function proj(lever, value) {
   const incre = value - (lever.lo ?? 0)
   const [ ha, sa, la ] = lever
   const hb = this.min >> 16 & 0x1FF, sb = this.min >> 8 & 0xFF, lb = this.min & 0xFF
-  return hsaToInt(scale(incre, ha, hb), limFF(incre, sa, sb), limFF(incre, la, lb))
+  return hsaToRgi(scale(incre, ha, hb), limFF(incre, sa, sb), limFF(incre, la, lb))
 }