+++ constants/enum-font-effects/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @palett/enum-font-effects
 
+## 0.12.6
+
+### Patch Changes
+
+- update algorithms form conversion and Pres
+
 ## 0.12.5
 
 ### Patch Changes