+++ packages/dye/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @palett/dye
 
+## 0.12.6
+
+### Patch Changes
+
+- update algorithms form conversion and Pres
+- Updated dependencies
+  - @palett/enum-ansi-codes@0.12.6
+  - @palett/enum-color-space@0.12.6
+  - @palett/enum-font-effects@0.12.6
+  - @palett/convert@0.12.6
+  - @palett/util-ansi@0.12.6
+
 ## 0.12.5
 
 ### Patch Changes