+++ packages/flopper/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @palett/flopper
 
+## 0.12.6
+
+### Patch Changes
+
+- update algorithms form conversion and Pres
+- Updated dependencies
+  - @palett/enum-color-space@0.12.6
+  - @palett/cards@0.12.6
+  - @palett/color-space@0.12.6
+  - @palett/convert@0.12.6
+  - @palett/munsell@0.12.6
+  - @palett/presets@0.12.6
+  - @palett/table@0.12.6
+
 ## 0.12.5
 
 ### Patch Changes