+++ packages/fluo/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @palett/fluo
 
+## 0.12.6
+
+### Patch Changes
+
+- update algorithms form conversion and Pres
+- Updated dependencies
+  - @palett/enum-ansi-codes@0.12.6
+  - @palett/dye@0.12.6
+  - @palett/presets@0.12.6
+  - @palett/projector@0.12.6
+  - @palett/util-ansi@0.12.6
+
 ## 0.12.5
 
 ### Patch Changes