+++ packages/projector/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @palett/projector
 
+## 0.12.6
+
+### Patch Changes
+
+- update algorithms form conversion and Pres
+- Updated dependencies
+  - @palett/dye@0.12.6
+  - @palett/presets@0.12.6
+
 ## 0.12.5
 
 ### Patch Changes