+++ packages/table/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @palett/table
 
+## 0.12.6
+
+### Patch Changes
+
+- update algorithms form conversion and Pres
+- Updated dependencies
+  - @palett/enum-color-cards@0.12.6
+  - @palett/enum-color-space@0.12.6
+  - @palett/enum-font-effects@0.12.6
+  - @palett/cards@0.12.6
+  - @palett/convert@0.12.6
+  - @palett/dye@0.12.6
+
 ## 0.12.5
 
 ### Patch Changes